# Conflicts:
#	package/commerce/src/Oro/Bundle/InventoryBundle/Resources/config/importexport.yml
#	package/commerce/src/Oro/Bundle/ProductBundle/Resources/views/Product/index.html.twig
#	package/platform/src/Oro/Bundle/ImportExportBundle/Tests/Behat/Context/ImportExportContext.php

+++ src/Oro/Bundle/UPSBundle/Handler/InvalidateCacheActionHandler.php
@@ -4,13 +4,20 @@ namespace Oro\Bundle\UPSBundle\Handler;
 
 use Oro\Bundle\CacheBundle\Action\Handler\InvalidateCacheActionHandlerInterface;
 use Oro\Bundle\CacheBundle\DataStorage\DataStorageInterface;
+use Oro\Bundle\EntityBundle\ORM\DoctrineHelper;
 use Oro\Bundle\ShippingBundle\Provider\Cache\ShippingPriceCache;
 use Oro\Bundle\UPSBundle\Cache\ShippingPriceCache as UPSShippingPriceCache;
 use Oro\Bundle\UPSBundle\TimeInTransit\CacheProvider\Factory\TimeInTransitCacheProviderFactoryInterface;
+use Oro\Bundle\UPSBundle\Entity\UPSTransport as UPSSettings;
 
 class InvalidateCacheActionHandler implements InvalidateCacheActionHandlerInterface
 {
     const PARAM_TRANSPORT_ID = 'transportId';
+
+    /**
+     * @var DoctrineHelper
+     */
+    private $doctrineHelper;
 
     /**
      * @var UPSShippingPriceCache
@@ -28,15 +35,18 @@ class InvalidateCacheActionHandler implements InvalidateCacheActionHandlerInterf
     private $timeInTransitCacheProviderFactory;
 
     /**
+     * @param DoctrineHelper                             $doctrineHelper
      * @param UPSShippingPriceCache                      $upsPriceCache
      * @param ShippingPriceCache                         $shippingPriceCache
      * @param TimeInTransitCacheProviderFactoryInterface $timeInTransitCacheProviderFactory
      */
     public function __construct(
+        DoctrineHelper $doctrineHelper,
         UPSShippingPriceCache $upsPriceCache,
         ShippingPriceCache $shippingPriceCache,
         TimeInTransitCacheProviderFactoryInterface $timeInTransitCacheProviderFactory
     ) {
+        $this->doctrineHelper = $doctrineHelper;
         $this->upsPriceCache = $upsPriceCache;
         $this->shippingPriceCache = $shippingPriceCache;
         $this->timeInTransitCacheProviderFactory = $timeInTransitCacheProviderFactory;
@@ -52,8 +62,22 @@ class InvalidateCacheActionHandler implements InvalidateCacheActionHandlerInterf
         $this->upsPriceCache->deleteAll($transportId);
         $this->shippingPriceCache->deleteAllPrices();
 
-        $this->timeInTransitCacheProviderFactory
-            ->createCacheProviderForTransport($transportId)
-            ->deleteAll();
+        $settings = $this->findSettings($transportId);
+
+        if ($settings !== null) {
+            $this->timeInTransitCacheProviderFactory
+                ->createCacheProviderForTransport($settings)
+                ->deleteAll();
+        }
+    }
+
+    /**
+     * @param int $settingsId
+     *
+     * @return UPSSettings|null|object
+     */
+    private function findSettings($settingsId)
+    {
+        return $this->doctrineHelper->getEntityRepository(UPSSettings::class)->find($settingsId);
     }
 }
